Some cleanup in the Datasources
