+++ src/main/java/io/frictionlessdata/tableschema/datasources/AbstractDataSource.java
@@ -1,15 +1,20 @@
 package io.frictionlessdata.tableschema.datasources;
 
+import com.google.common.collect.Iterators;
 import org.apache.commons.csv.CSVFormat;
+import org.apache.commons.csv.CSVParser;
 import org.apache.commons.csv.CSVPrinter;
+import org.apache.commons.csv.CSVRecord;
 import org.json.JSONArray;
 
 import java.io.*;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
+import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
@@ -44,17 +49,95 @@ public abstract class AbstractDataSource implements DataSource {
         this.dataSource = dataSource;
     }
 
+    abstract CSVParser getCSVParser() throws Exception;
+
     @Override
-    abstract public Iterator<String[]> iterator() throws Exception;
-    
+    public Iterator<String[]> iterator() throws Exception{
+        Iterator<CSVRecord> iterCSVRecords = this.getCSVParser().iterator();
+
+        Iterator<String[]> iterStringArrays = Iterators.transform(iterCSVRecords, (CSVRecord input) -> {
+            Iterator<String> iterCols = input.iterator();
+
+            List<String> cols = new ArrayList();
+            while(iterCols.hasNext()){
+                cols.add(iterCols.next());
+            }
+
+            String[] output = cols.toArray(new String[0]);
+
+            return output;
+        });
+
+        return iterStringArrays;
+    }
+
     @Override
-    abstract public String[] getHeaders() throws Exception;
-    
+    public List<String[]> data() throws Exception{
+        // This is pretty much what happens when we call this.parser.getRecords()...
+        Iterator<CSVRecord> iter = this.getCSVParser().iterator();
+        List<String[]> data = new ArrayList();
+
+        while(iter.hasNext()){
+            CSVRecord record = iter.next();
+            Iterator<String> colIter = record.iterator();
+
+            //...except that we want list of String[] rather than list of CSVRecord.
+            List<String> cols = new ArrayList();
+            while(colIter.hasNext()){
+                cols.add(colIter.next());
+            }
+
+            data.add(cols.toArray(new String[0]));
+        }
+
+        return data;
+    }
+
     @Override
-    abstract public List<String[]> data() throws Exception;
-    
+    public void write(File outputFile) throws Exception {
+        try (Writer out = new BufferedWriter(new FileWriter(outputFile));
+             CSVPrinter csvPrinter = new CSVPrinter(out, CSVFormat.RFC4180)) {
+
+            if (this.getHeaders() != null) {
+                csvPrinter.printRecord(this.getHeaders());
+            }
+
+            Iterator<CSVRecord> recordIter = this.getCSVParser().iterator();
+            while(recordIter.hasNext()){
+                CSVRecord record = recordIter.next();
+                csvPrinter.printRecord(record);
+            }
+
+            csvPrinter.flush();
+
+        } catch (Exception e) {
+            throw e;
+        }
+    }
+
     @Override
-    abstract public void write(File outputFile) throws Exception;
+    public String[] getHeaders() throws Exception{
+        try{
+            // Get a copy of the header map that iterates in column order.
+            // The map keys are column names. The map values are 0-based indices.
+            Map<String, Integer> headerMap = this.getCSVParser().getHeaderMap();
+
+            // Generate list of keys
+            List<String> headerVals = new ArrayList();
+
+            headerMap.entrySet().forEach((pair) -> {
+                headerVals.add((String)pair.getKey());
+            });
+
+            // Return string array of keys.
+            return headerVals.toArray(new String[0]);
+
+        }catch(Exception e){
+            throw e;
+        }
+
+    }
+
 
     String getFileContents(String path) throws IOException {
         String lines;
